+++ processor/src/test/resources/bad_input/com/example/computed_value_ref/BadReturnTypeModel.java
@@ -1,8 +1,8 @@
 package com.example.computed_value_ref;
 
 import arez.annotations.ArezComponent;
+import arez.annotations.ComputableValueRef;
 import arez.annotations.Computed;
-import arez.annotations.ComputedValueRef;
 import javax.annotation.Nonnull;
 
 @ArezComponent
@@ -15,6 +15,6 @@ public abstract class BadReturnTypeModel
   }
 
   @Nonnull
-  @ComputedValueRef
+  @ComputableValueRef
   abstract String getTimeComputableValue();
 }