+++ processor/src/test/resources/input/com/example/computable_value_ref/NonStandardNameModel.java
@@ -1,9 +1,9 @@
-package com.example.computed_value_ref;
+package com.example.computable_value_ref;
 
 import arez.ComputableValue;
 import arez.annotations.ArezComponent;
+import arez.annotations.ComputableValueRef;
 import arez.annotations.Computed;
-import arez.annotations.ComputedValueRef;
 
 @ArezComponent
 public abstract class NonStandardNameModel
@@ -14,6 +14,6 @@ public abstract class NonStandardNameModel
     return 0;
   }
 
-  @ComputedValueRef( name = "time" )
+  @ComputableValueRef( name = "time" )
   abstract ComputableValue timeComputableValue();
 }