+++ processor/src/test/resources/input/com/example/computable_value_ref/RawComputableValueModel.java
@@ -1,12 +1,12 @@
-package com.example.computed_value_ref;
+package com.example.computable_value_ref;
 
 import arez.ComputableValue;
 import arez.annotations.ArezComponent;
 import arez.annotations.Computed;
-import arez.annotations.ComputedValueRef;
+import arez.annotations.ComputableValueRef;
 
 @ArezComponent
-public abstract class RawComputedValueModel
+public abstract class RawComputableValueModel
 {
   @Computed
   public long getTime()
@@ -14,6 +14,6 @@ public abstract class RawComputedValueModel
     return 0;
   }
 
-  @ComputedValueRef
+  @ComputableValueRef
   abstract ComputableValue getTimeComputableValue();
 }